--- src/components/Cart/Module/contants/contant.js
@@ -1,4 +1,0 @@
-export const ADD_TO_CARD="ADD_TO_CARD";
-export const REMOVE_CARD="REMOVE_CARD";
-export const POST_CART="POST_CART";
-export const IS_LOADING="IS_LOADING";
