+++ src/components/Cart/module/Contants/Contants.jsx
@@ -1,0 +1,8 @@
+export const ADD_TO_CARD = "ADD_TO_CARD"
+export const REMOVE_TO_CARD = "REMOVE_TO_CARD"
+export const UPDATE_SIZE_COLOR  = "UPDATE_SIZE_COLOR"
+export const RESET_CART  = "RESET_CART"
+
+export const ADD_TO_CARDFAVOR = "ADD_TO_CARDFAVOR"
+export const REMOVECART_TOFAVOR = "REMOVECART_TOFAVOR"
+export const DELETE_FAVOR = "DELETE_FAVOR"
