+++ src/components/List Product/module/Constants/contants.js
@@ -1,0 +1,10 @@
+export const CHANGE_GENDER_TYPEPRODUCT = "CHANGE_GENDER_TYPEPRODUCT"
+export const FETCH_API_LISTPRODUCT = "FETCH_API_LISTPRODUCT"
+export const SORT_DATA = "SORT_DATA"
+export const FILTER_COLOR = "FILTER_COLOR"
+export const FILTER_SIZE = "FILTER_SIZE"
+export const FILTER_COLOR_DATA = "FILTER_COLOR_DATA"
+export const SORT_BY_TITLE = "SORT_BY_TITLE"
+export const DATA_SEARCH = "DATA_SEARCH"
+export const DATA_SEARCH_INPUT = "DATA_SEARCH_INPUT"
+export const IS_LOADING_LIST_PRODUCT = "IS_LOADING_LIST_PRODUCT"
