--- src/components/ListProductComponent/module/contans/contants.js
@@ -1,9 +1,0 @@
-export const IS_LOADING="IS_LOADING";
-export const GET_DATA="GET_DATA";
-export const CHANGER_TYPEPRODUCT_GENDER="CHANGER_TYPEPRODUCT_GENDER";
-export const FILTER_DATA="FILTER_DATA";
-export const FILTER_COLOR="FILTER_COLOR";
-export const FILTER_SIZE="FILTER_SIZE";
-export const SORT_TITLE="SORT_TITLE";
-export const DATA_SEARCH="DATA_SEARCH";
-export const DATA_SEARCH_INPUT="DATA_SEARCH_INPUT";
