+++ src/components/NavBar/NavMainComponents/Redux/Modules/constants/Constants.js
@@ -1,0 +1,9 @@
+export const REDUX = "OPEN_SIGNIN";
+export const SIGNUP = "OPEN_SIGNUP";
+export const FETCH_API_LOGIN = "FETCH_API_LOGIN";
+export const LOGOUT = "LOG_OUT";
+export const DATA_ALL = "DATA_ALL";
+export const SEARCH = "SEARCH";
+export const SUGGEST = "SUGGEST";
+export const SET_ADMIN = "SET_ADMIN";
+
