+++ src/redux/rootReducers.js
@@ -1,12 +1,12 @@
 import { combineReducers } from "redux";
 import reducer from "./reducers/productReducers"
-import listProductReducer from "../components/ListProductComponent/module/reducers/listProductReducer"
-import searchBarReducer from  "../components/NavBar/module/reducers/SearchBarReducer";
-import cartReducer from "../components/Cart/Module/reducer/CartReducer"
+import listProductReducer from "../components/List Product/module/Reducers/reducer"
+// import searchBarReducer from  "../components/NavBar/module/reducers/SearchBarReducer";
+import reducerCart from "../components/Cart/module/Reducers/reducer"
 const rootReducers = combineReducers({
     reducer,
     listProductReducer,
-    searchBarReducer,
-    cartReducer,
+    // searchBarReducer,
+    reducerCart,
 })
 export default rootReducers;
